gateway/requirements.txt
fastapi
"uvicorn[standard]"
psycopg2-binary
python-dotenv
sqlmodel
SQLAlchemy-Utils
pika
sqlalchemy-json
python-multipart